V4 Removes restriction  =on number of tests
This version uses the SDcard storage to remove the original restriction imposed by the lack of memory on the arduino nano.  It not super pretty but works well. the data is left persession as a csv file on the sd card.  The card may eventually fill up but both the data and log files written there are tiny.